backend/requirements.txt
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
Pillow==8.2.0
